Add files via upload
The Instagram Follower Bot is a simple JavaScript script designed to automate the process of following a list of users on Instagram. The script uses the Instagram API to access and interact with Instagram accounts. 

How it works: To use the bot, the user needs to enter the "following" section of their Instagram profile and run the script in their browser's console. The script will then start following the list of usernames provided by the user. It works by using the Instagram API to interact with the accounts, and by sending automated follow requests to each account on the list. 

The Benefits: The Instagram Follower Bot provides several benefits for users who want to be seen on Instagram in some way. First and foremost, it saves time and effort by automating the process of following accounts, allowing users to focus on creating content and engaging with their followers. Additionally, the bot can help increase engagement and reach by following users who are likely to be interested in the user's content. By following accounts that are relevant to the user's niche or audience, the bot can attract more followers who are genuinely interested in the user's content. Follow me on Instagram: If you have any questions or would like to learn more about the Instagram Follower Bot, you can follow me on Instagram at @vinicius.acacioo. I am always available to answer any questions or provide support for users who are interested in using the bot to grow their Instagram following.